Add CouchbaseSearchVectorStore and refactor vector store structure
- Introduced CouchbaseSearchVectorStore for enhanced search capabilities.
- Refactored vector store imports and structure for better organization.
- Added base vector store functionality with comprehensive methods for document management and search.
- Implemented integration tests to validate the functionality of the new search vector store.
- Updated __init__.py to include new classes in the module exports.
