+++ langchain_couchbase/__init__.py
@@ -1,8 +1,12 @@
 from langchain_couchbase.cache import CouchbaseCache, CouchbaseSemanticCache
 from langchain_couchbase.chat_message_histories import CouchbaseChatMessageHistory
-from langchain_couchbase.vectorstores import CouchbaseVectorStore
+from langchain_couchbase.vectorstores import (
+    CouchbaseSearchVectorStore,
+    CouchbaseVectorStore,
+)
 
 __all__ = [
+    "CouchbaseSearchVectorStore",
     "CouchbaseVectorStore",
     "CouchbaseCache",
     "CouchbaseSemanticCache",
