+++ langchain_couchbase/vectorstores/__init__.py
@@ -1,0 +1,9 @@
+from langchain_couchbase.vectorstores.search_vector_store import (
+    CouchbaseSearchVectorStore,
+)
+from langchain_couchbase.vectorstores.vectorstores import CouchbaseVectorStore
+
+__all__ = [
+    "CouchbaseSearchVectorStore",
+    "CouchbaseVectorStore",
+]
